+++ src/environment.h
@@ -32,7 +32,7 @@ public:
     Environment(const LONG client_width, const LONG client_height);
     void Render(const size_t iteration, const size_t action, float exploration_rate);
     size_t Reset();
-    std::tuple<size_t, int, bool> Step(const size_t action);
+    std::tuple<size_t, float, bool> Step(const size_t action);
 
     LONG client_width, client_height;
     LONG minLeft = 0;
@@ -64,7 +64,7 @@ private:
     std::string thirstStateStr;
     std::string hungerStateStr;
     std::string actionStr;
-    int reward;
+    float reward;
     size_t daysLived;
     size_t daysWithoutEating;
     size_t daysWithoutDrinking;
