+++ src/q_learning.h
@@ -7,7 +7,7 @@ class QLearning
 public:
     QLearning(size_t n_states, size_t n_actions, float learning_rate = 0.01f, float discount_factor = 0.5f, float exploration_rate = 1.0f, float exploration_decay = 0.995f, float exploration_min = 0.01f);
     size_t ChooseAction(size_t state);
-    void UpdateQtable(size_t state, size_t action, int reward, size_t next_state, bool done);
+    void UpdateQtable(size_t state, size_t action, float reward, size_t next_state, bool done);
     
     float exploration_rate;
